+++ src/msl/io/tables.py
@@ -14,6 +14,7 @@ from .readers import ExcelReader, GSheetsReader, ODSReader
 from .utils import get_basename, get_extension, get_lines
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
     from datetime import date, datetime
     from typing import Any
 
@@ -149,6 +150,7 @@ def read_table_excel(  # noqa: PLR0913
     as_datetime: bool = True,
     dtype: DTypeLike | None = None,
     invalid_date: str | date | datetime | None = None,
+    skip_rows: Iterable[int] | None = None,
     **kwargs: Any,
 ) -> Dataset:
     """Read a data table from an Excel spreadsheet.
@@ -172,6 +174,7 @@ def read_table_excel(  # noqa: PLR0913
         invalid_date: If `None`, an error is raised if a cell contains a value that
             is an invalid date. If not `None`, all cells that contain an invalid date are
             replaced with the specified value.
+        skip_rows: Row numbers to skip. The first row number in a spreadsheet is 1 (not 0).
         kwargs: All additional keyword arguments are passed to [xlrd.open_workbook][].
             Can use an `encoding` keyword argument as an alias for `encoding_override`.
 
@@ -187,7 +190,7 @@ def read_table_excel(  # noqa: PLR0913
             num_rows, num_cols = excel.dimensions(name)
             letters = excel.to_letters(num_cols - 1)
             cells += f":{letters}{num_rows}"
-        table = excel.read(cells, sheet=sheet, as_datetime=as_datetime, invalid_date=invalid_date)
+        table = excel.read(cells, sheet=sheet, as_datetime=as_datetime, invalid_date=invalid_date, skip_rows=skip_rows)
 
     return _spreadsheet_to_dataset(table, file, dtype)
 
@@ -200,6 +203,7 @@ def read_table_gsheets(  # noqa: PLR0913
     as_datetime: bool = True,
     dtype: DTypeLike | None = None,
     invalid_date: datetime | None = None,
+    skip_rows: Iterable[int] | None = None,
     **kwargs: Any,
 ) -> Dataset:
     """Read a data table from a Google Sheets spreadsheet.
@@ -227,6 +231,7 @@ def read_table_gsheets(  # noqa: PLR0913
         invalid_date: If `None`, an error is raised if a cell contains a value that
             is an invalid date. If not `None`, all cells that contain an invalid date are
             replaced with the specified value.
+        skip_rows: Row numbers to skip. The first row number in a spreadsheet is 1 (not 0).
         kwargs: All additional keyword arguments are passed to [GSheetsReader][msl.io.readers.gsheets.GSheetsReader].
 
     Returns:
@@ -242,10 +247,12 @@ def read_table_gsheets(  # noqa: PLR0913
                 raise ValueError(msg)
 
             r, c = sheets.to_indices(cells)
-            data = sheets.read(sheet=sheet, as_datetime=as_datetime, invalid_date=invalid_date)
+            data = sheets.read(sheet=sheet, as_datetime=as_datetime, invalid_date=invalid_date, skip_rows=skip_rows)
             table = [row[c:] for row in data[r:]]
         else:
-            table = sheets.read(cells, sheet=sheet, as_datetime=as_datetime, invalid_date=invalid_date)
+            table = sheets.read(
+                cells, sheet=sheet, as_datetime=as_datetime, invalid_date=invalid_date, skip_rows=skip_rows
+            )
 
     return _spreadsheet_to_dataset(table, file, dtype)
 
@@ -256,9 +263,10 @@ def read_table_ods(  # noqa: PLR0913
     cells: str | None = None,
     sheet: str | None = None,
     as_datetime: bool = True,
-    merged: bool = False,
     dtype: DTypeLike | None = None,
     invalid_date: str | date | datetime | None = None,
+    merged: bool = False,
+    skip_rows: Iterable[int] | None = None,
     **kwargs: Any,
 ) -> Dataset:
     """Read a data table from an OpenDocument Spreadsheet.
@@ -278,15 +286,16 @@ def read_table_ods(  # noqa: PLR0913
         as_datetime: Whether dates should be returned as [datetime][datetime.datetime] or
             [date][datetime.date] objects. If `False`, dates are returned as a string in
             the display format of the spreadsheet cell.
+        dtype: The data type(s) to use for the table.
+        invalid_date: If `None`, an error is raised if a cell contains a value that
+            is an invalid date. If not `None`, all cells that contain an invalid date are
+            replaced with the specified value.
         merged: Applies to cells that are merged with other cells. If `False`, the
             value of each unmerged cell is returned, otherwise the same value is
             returned for all merged cells. In an OpenDocument Spreadsheet, the value
             of a hidden cell that is merged with a visible cell can still be retained
             (depends on how the merger was performed).
-        dtype: The data type(s) to use for the table.
-        invalid_date: If `None`, an error is raised if a cell contains a value that
-            is an invalid date. If not `None`, all cells that contain an invalid date are
-            replaced with the specified value.
+        skip_rows: Row numbers to skip. The first row number in a spreadsheet is 1 (not 0).
         kwargs: All keyword arguments are passed to [ODSReader][msl.io.readers.ods.ODSReader].
 
     Returns:
@@ -300,7 +309,9 @@ def read_table_ods(  # noqa: PLR0913
             num_rows, num_cols = ods.dimensions(name)
             letters = ods.to_letters(num_cols - 1)
             cells += f":{letters}{num_rows}"
-        table = ods.read(cells, sheet=sheet, as_datetime=as_datetime, merged=merged, invalid_date=invalid_date)
+        table = ods.read(
+            cells, sheet=sheet, as_datetime=as_datetime, merged=merged, invalid_date=invalid_date, skip_rows=skip_rows
+        )
     return _spreadsheet_to_dataset(table, file, dtype)